Create index.md
f or relating to the human soul or mind; mental (opposed to physical).
Psychology. pertaining to or noting mental phenomena.
outside of natural or scientific knowledge; spiritual.
of or relating to some apparently nonphysical force or agency:
psychic research; psychic phenomena.
sensitive to influences or forces of a nonphysical or supernatural nature.